requirements.txt
flask
gradio
groq
python-dotenv